apps/backend/requirements.txt
# MathVerse Backend API - Core Dependencies
# This file contains all Python dependencies for the backend API service

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
saffier==2.5.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Data Serialization
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic[email]==2.5.3
orjson==3.9.15

# Database and ORM
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Configuration Management
pyyaml==6.0.1
python-dotenv==1.0.0

# Async Support
httpx==0.26.0
aiofiles==23.2.1

# Email Services
emails==0.6.8
aiosmtplib==3.0.1

# Caching
redis==5.0.1
cachetools==5.3.2

# File Storage
aiobotocore==2.7.0
aiofiles==23.2.1

# Rate Limiting
slowapi==0.1.9

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-httpx==0.28.0
pytest-timeout==2.3.1

# Code Quality
black==23.12.1
isort==5.13.2
pylint==3.0.3
mypy==1.8.0

# Development Utilities
watchdog==4.0.0
ipython==8.21.0
